+++ ubo_app/services/000-sound/ubo_handle.py
@@ -12,7 +12,6 @@ def setup(service: Service) -> None:
     from setup import init_service
 
     service.register_reducer(reducer)
-
     init_service()