+++ ubo_app/store/__init__.py
@@ -31,6 +31,7 @@ from ubo_app.store.services.rgb_ring import RgbRingAction
 from ubo_app.store.services.sensors import SensorsAction, SensorsState
 from ubo_app.store.services.sound import SoundAction, SoundState
 from ubo_app.store.services.ssh import SSHAction, SSHState
+from ubo_app.store.services.voice import VoiceAction, VoiceState
 from ubo_app.store.services.wifi import WiFiAction, WiFiEvent, WiFiState
 from ubo_app.store.status_icons import StatusIconsAction, StatusIconsState
 from ubo_app.store.status_icons.reducer import reducer as status_icons_reducer
@@ -69,6 +70,7 @@ class RootState(BaseCombineReducerState):
     ip: IpState
     notifications: NotificationsState
     docker: DockerState
+    voice: VoiceState
 
 
 class ScreenshotEvent(BaseEvent): ...
@@ -89,6 +91,7 @@ ActionType = (
     | NotificationsAction
     | DockerAction
     | RgbRingAction
+    | VoiceAction
 )
 EventType = KeypadEvent | CameraEvent | WiFiEvent | IpEvent | ScreenshotEvent