+++ ubo_app/store/services/voice.py
@@ -1,0 +1,24 @@
+# ruff: noqa: D100, D101, D102, D103, D104, D107, N999
+from __future__ import annotations
+
+from immutable import Immutable
+from redux import BaseAction, BaseEvent
+
+
+class VoiceAction(BaseAction): ...
+
+
+class VoiceEvent(BaseEvent): ...
+
+
+class VoiceReadTextAction(VoiceAction):
+    text: str
+    speech_rate: float | None = None
+
+
+class VoiceSynthesizeTextEvent(VoiceEvent):
+    text: str
+    speech_rate: float | None = None
+
+
+class VoiceState(Immutable): ...